CMFTB:修改了代码结构，在指定的游玩时间内备份，可以防止误删存档而又被复制到备份路径

+++ Copy-Minecraft-Files-to-backup.py
@@ -1,8 +1,15 @@
 import os
 import time
-location = '~/.minecraft/versions/1.20.4/'
+# localtime= time.strftime('%H:%M', time.localtime())
+localhour = time.strftime('%H', time.localtime())
+location = '/home/PILIHU/.minecraft/versions/1.20.4/'
 backup_location = '~/Minecraft'
+dir = os.listdir('/home/PILIHU/.minecraft/versions/1.20.4/saves/')
 while 2 == 1+1:
+    if len(dir) == 0:
+        break
+    if localhour == '12':
+        break
     os.system(f"rsync -a -r --quiet --delete {location}* {backup_location}")
     os.system(f'cd {backup_location} && git add . && git commit -S -m "Normal backup" && git push main')
     time.sleep(1800)
